// projects/shared/src/lib/button/button.enum.ts
export enum ButtonStatus {
    SUBMIT = 'Submit',
    WORKING = 'Working...',
    FAILED = 'Failed',
    SUCCESS = 'Success!',
    OUTLINESUBMIT = 'OutlineSubmit(S)',
    OUTLINEFAILED = 'OutlineFailed',
    OUTLINESUCCESS = 'OutlineSuccess!',
    OUTLINEWORKING = 'OutlineWorking...',
    DISABLED = 'disabled'
}